Move graph_greedy_coloring to its own file
